Added checks for certain API-keys to not copy or reset

+++ v2/models/copier.js
@@ -15,51 +15,34 @@ const copier = {
     copyApiKey: process.env.COPY_API_KEY || config.copyApiKey,
 
     copyAll: function(res, apiKey) {
-        let sql = "INSERT INTO products" +
-            " (articleNumber," +
-            " productName," +
-            " productDescription," +
-            " productSpecifiers," +
-            " stock," +
-            " location," +
-            " price," +
-            " apiKey)" +
-            " SELECT articleNumber," +
-            " productName," +
-            " productDescription," +
-            " productSpecifiers," +
-            " stock," +
-            " location, " +
-            " price," +
-            "'" + apiKey + "'" +
-            " FROM products" +
-            " WHERE apiKey = ?";
+        if (apiKey === copier.copyApiKey) {
+            var err = { message: "Cannot copy that API-key." };
 
-        db.run(sql,
-            copier.copyApiKey,
-            function (err) {
-                if (err) {
-                    return res.status(500).json({
-                        errors: {
-                            status: 500,
-                            source: "/copy_products",
-                            title: "Database error",
-                            detail: err.message
-                        }
-                    });
-                }
+            return copier.errorResponse(res, "/v2/copier/reset", err);
+        } else {
+            let sql = "INSERT INTO products" +
+                " (articleNumber," +
+                " productName," +
+                " productDescription," +
+                " productSpecifiers," +
+                " stock," +
+                " location," +
+                " price," +
+                " apiKey)" +
+                " SELECT articleNumber," +
+                " productName," +
+                " productDescription," +
+                " productSpecifiers," +
+                " stock," +
+                " location, " +
+                " price," +
+                "'" + apiKey + "'" +
+                " FROM products" +
+                " WHERE apiKey = ?";
 
-                let productChanges = this.changes;
-                let productLast = this.lastID;
-
-                db.all("SELECT * FROM " +
-                    "(SELECT " + products.dataFields +
-                    " FROM products WHERE apiKey = ?" +
-                    " ORDER BY ROWID DESC LIMIT " +
-                    parseInt(this.changes) + ")" +
-                    " ORDER BY id ASC",
-                apiKey,
-                (err, rows) => {
+            db.run(sql,
+                copier.copyApiKey,
+                function (err) {
                     if (err) {
                         return res.status(500).json({
                             errors: {
@@ -71,134 +54,143 @@ const copier = {
                         });
                     }
 
-                    let copiedProducts = rows;
+                    let productChanges = this.changes;
+                    let productLast = this.lastID;
 
-                    let sql = "INSERT INTO orders" +
-                        " (customerName," +
-                        " customerAddress," +
-                        " customerZip," +
-                        " customerCity," +
-                        " customerCountry," +
-                        " statusId," +
-                        " apiKey)" +
-                        " SELECT customerName," +
-                        " customerAddress," +
-                        " customerZip," +
-                        " customerCity," +
-                        " customerCountry," +
-                        " statusId," +
-                        "'" + apiKey + "'" +
-                        " FROM orders" +
-                        " WHERE apiKey = ?";
-
-                    db.run(sql,
-                        copier.copyApiKey,
-                        function (err) {
-                            if (err) {
-                                return res.status(500).json({
-                                    errors: {
-                                        status: 500,
-                                        source: "/copy_orders",
-                                        title: "Database error",
-                                        detail: err.message
-                                    }
-                                });
-                            }
-
-                            let copyResponse = {
-                                data: {
-                                    products: copiedProducts,
-                                    orders: []
+                    db.all("SELECT * FROM " +
+                        "(SELECT " + products.dataFields +
+                        " FROM products WHERE apiKey = ?" +
+                        " ORDER BY ROWID DESC LIMIT " +
+                        parseInt(this.changes) + ")" +
+                        " ORDER BY id ASC",
+                    apiKey,
+                    (err, rows) => {
+                        if (err) {
+                            return res.status(500).json({
+                                errors: {
+                                    status: 500,
+                                    source: "/copy_products",
+                                    title: "Database error",
+                                    detail: err.message
                                 }
-                            };
+                            });
+                        }
 
-                            let productOffset = productLast - productChanges;
-                            let orderOffset = this.lastID - this.changes;
+                        let copiedProducts = rows;
 
-                            let range = {
-                                min: orderOffset + 1,
-                                max: this.lastID,
-                            };
+                        let sql = "INSERT INTO orders" +
+                            " (customerName," +
+                            " customerAddress," +
+                            " customerZip," +
+                            " customerCity," +
+                            " customerCountry," +
+                            " statusId," +
+                            " apiKey)" +
+                            " SELECT customerName," +
+                            " customerAddress," +
+                            " customerZip," +
+                            " customerCity," +
+                            " customerCountry," +
+                            " statusId," +
+                            "'" + apiKey + "'" +
+                            " FROM orders" +
+                            " WHERE apiKey = ?";
 
-                            let oisql = `INSERT INTO order_items
-                                (orderId, productId, amount, apiKey)
-                                SELECT (orderId + ${orderOffset}),
-                                (productId + ${productOffset}),
-                                amount,
-                                '${apiKey}'
-                                FROM order_items WHERE apiKey = ?`;
+                        db.run(sql,
+                            copier.copyApiKey,
+                            function (err) {
+                                if (err) {
+                                    return res.status(500).json({
+                                        errors: {
+                                            status: 500,
+                                            source: "/copy_orders",
+                                            title: "Database error",
+                                            detail: err.message
+                                        }
+                                    });
+                                }
 
-                            db.run(
-                                oisql,
-                                copier.copyApiKey,
-                                function (err) {
-                                    if (err) {
-                                        return res.status(500).json({
-                                            errors: {
-                                                status: 500,
-                                                source: "/copy_orders",
-                                                title: "Database error",
-                                                detail: err.message
-                                            }
-                                        });
+                                let copyResponse = {
+                                    data: {
+                                        products: copiedProducts,
+                                        orders: []
                                     }
+                                };
 
-                                    return orders.getAllOrders(
-                                        res,
-                                        apiKey,
-                                        201,
-                                        copyResponse,
-                                        range,
-                                    );
-                                });
-                        });
+                                let productOffset = productLast - productChanges;
+                                let orderOffset = this.lastID - this.changes;
+
+                                let range = {
+                                    min: orderOffset + 1,
+                                    max: this.lastID,
+                                };
+
+                                let oisql = `INSERT INTO order_items
+                                    (orderId, productId, amount, apiKey)
+                                    SELECT (orderId + ${orderOffset}),
+                                    (productId + ${productOffset}),
+                                    amount,
+                                    '${apiKey}'
+                                    FROM order_items WHERE apiKey = ?`;
+
+                                db.run(
+                                    oisql,
+                                    copier.copyApiKey,
+                                    function (err) {
+                                        if (err) {
+                                            return res.status(500).json({
+                                                errors: {
+                                                    status: 500,
+                                                    source: "/copy_orders",
+                                                    title: "Database error",
+                                                    detail: err.message
+                                                }
+                                            });
+                                        }
+
+                                        return orders.getAllOrders(
+                                            res,
+                                            apiKey,
+                                            201,
+                                            copyResponse,
+                                            range,
+                                        );
+                                    });
+                            });
+                    });
                 });
-            });
+        }
     },
 
     copyProducts: function(res, apiKey) {
-        let sql = "INSERT INTO products" +
-            " (articleNumber," +
-            " productName," +
-            " productDescription," +
-            " productSpecifiers," +
-            " stock," +
-            " location," +
-            " price," +
-            " apiKey)" +
-            " SELECT articleNumber," +
-            " productName," +
-            " productDescription," +
-            " productSpecifiers," +
-            " stock," +
-            " location, " +
-            " price," +
-            " '" + apiKey + "'" +
-            " FROM products" +
-            " WHERE apiKey = ?";
+        if (apiKey === copier.copyApiKey) {
+            var err = { message: "Cannot copy that API-key." };
 
-        db.run(sql,
-            copier.copyApiKey,
-            function (err) {
-                if (err) {
-                    return res.status(500).json({
-                        errors: {
-                            status: 500,
-                            source: "/copy_products",
-                            title: "Database error",
-                            detail: err.message
-                        }
-                    });
-                }
+            return copier.errorResponse(res, "/v2/copier/reset", err);
+        } else {
+            let sql = "INSERT INTO products" +
+                " (articleNumber," +
+                " productName," +
+                " productDescription," +
+                " productSpecifiers," +
+                " stock," +
+                " location," +
+                " price," +
+                " apiKey)" +
+                " SELECT articleNumber," +
+                " productName," +
+                " productDescription," +
+                " productSpecifiers," +
+                " stock," +
+                " location, " +
+                " price," +
+                " '" + apiKey + "'" +
+                " FROM products" +
+                " WHERE apiKey = ?";
 
-                db.all("SELECT * FROM " +
-                    "(SELECT " + products.dataFields +
-                    " FROM products WHERE apiKey = ?" +
-                    " ORDER BY ROWID DESC LIMIT " +
-                    parseInt(this.changes) + ")" +
-                    " ORDER BY id ASC",
-                apiKey,
-                (err, rows) => {
+            db.run(sql,
+                copier.copyApiKey,
+                function (err) {
                     if (err) {
                         return res.status(500).json({
                             errors: {
@@ -210,49 +202,75 @@ const copier = {
                         });
                     }
 
-                    res.status(201).json( { data: rows } );
+                    db.all("SELECT * FROM " +
+                        "(SELECT " + products.dataFields +
+                        " FROM products WHERE apiKey = ?" +
+                        " ORDER BY ROWID DESC LIMIT " +
+                        parseInt(this.changes) + ")" +
+                        " ORDER BY id ASC",
+                    apiKey,
+                    (err, rows) => {
+                        if (err) {
+                            return res.status(500).json({
+                                errors: {
+                                    status: 500,
+                                    source: "/copy_products",
+                                    title: "Database error",
+                                    detail: err.message
+                                }
+                            });
+                        }
+
+                        res.status(201).json( { data: rows } );
+                    });
                 });
-            });
+        }
     },
 
     reset: function(res, apiKey) {
-        db.run("DELETE FROM products WHERE apiKey=?", apiKey, (err) => {
-            if (err) {
-                return copier.errorResponse(res, "/v2/copier/reset", err);
-            }
+        if (apiKey === copier.copyApiKey) {
+            var err = { message: "Cannot reset that API-key." };
 
-            db.run("DELETE FROM orders WHERE apiKey=?", apiKey, (err) => {
+            return copier.errorResponse(res, "/v2/copier/reset", err);
+        } else {
+            db.run("DELETE FROM products WHERE apiKey=?", apiKey, (err) => {
                 if (err) {
                     return copier.errorResponse(res, "/v2/copier/reset", err);
                 }
 
-                db.run("DELETE FROM order_items WHERE apiKey=?", apiKey, (err) => {
+                db.run("DELETE FROM orders WHERE apiKey=?", apiKey, (err) => {
                     if (err) {
                         return copier.errorResponse(res, "/v2/copier/reset", err);
                     }
 
-                    db.run("DELETE FROM deliveries WHERE apiKey=?", apiKey, (err) => {
+                    db.run("DELETE FROM order_items WHERE apiKey=?", apiKey, (err) => {
                         if (err) {
                             return copier.errorResponse(res, "/v2/copier/reset", err);
                         }
 
-                        db.run("DELETE FROM invoices WHERE apiKey=?", apiKey, (err) => {
+                        db.run("DELETE FROM deliveries WHERE apiKey=?", apiKey, (err) => {
                             if (err) {
                                 return copier.errorResponse(res, "/v2/copier/reset", err);
                             }
 
-                            db.run("DELETE FROM users WHERE apiKey=?", apiKey, (err) => {
+                            db.run("DELETE FROM invoices WHERE apiKey=?", apiKey, (err) => {
                                 if (err) {
                                     return copier.errorResponse(res, "/v2/copier/reset", err);
                                 }
 
-                                return copier.copyAll(res, apiKey);
+                                db.run("DELETE FROM users WHERE apiKey=?", apiKey, (err) => {
+                                    if (err) {
+                                        return copier.errorResponse(res, "/v2/copier/reset", err);
+                                    }
+
+                                    return copier.copyAll(res, apiKey);
+                                });
                             });
                         });
                     });
                 });
             });
-        });
+        }
     },
 
     errorResponse: function(res, path, err) {
